Updating the SB generator from 0.0.14 to 0.0.19.
Additionally, updating NS version to 2.2.0.

After updating SB generator, it made some modifications to other files. I believe that these modifications are not going to break the tutorial logic. I'm not including homeView-service.js and homeView-view-model.js files in the PR as the users are going to create them during the tutorial.